+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Release 1.3.0
+=============
+
+- additional-header parameter `-header`
+- additional-header-from-file parameter `-headerFile`
+
 Release 1.2.1
 =============